+++ era5cli/args/__init__.py
@@ -1,0 +1,7 @@
+from era5cli.args import common
+from era5cli.args import config
+from era5cli.args import info
+from era5cli.args import periods
+
+
+__all__ = ["common", "config", "periods", "info"]